+++ src/Network/Serialize/Verifications/ParseMemoryVerification.nim
@@ -37,7 +37,7 @@ proc parseMemoryVerification*(
         #Verifier's Public Key.
         verifier: BLSPublicKey
         #Nonce.
-        nonce: uint = uint(verifSeq[1].fromBinary())
+        nonce: int = verifSeq[1].fromBinary()
         #Get the Entry hash.
         entry: string = verifSeq[2]
         #BLS signature.
